+++ src/widgets/ruleswidget.hpp
@@ -46,7 +46,7 @@ class RulesWidget : public QWidget
 
     signals:
         void gameInfoChangedSignal(const QString& gameInfo);
-        void setMaxIdleTimeSignal();
+        void refreshAFKTimersSignal();
         void setMaxPlayersSignal(const qint32& maxPlayers);
 
     public slots: